Craeted shape.py and did some refactoring
